Get Zengine API data

get_zen() gets data from the Zengine API.

get_zen(resource, id = NULL, record = NULL, limit = NULL)

Arguments

resource
A character vector specifying a Zengine API resource. Common choices include "forms" and "workspaces". A full list of resources is available here. Defaults to "forms".

id
A character vector specifying a resource ID. Defaults to NULL.

record
A character or numeric vector specifying a record linked to a resource ID. The id argument must be provided if record is used. Defaults to NULL.

limit
An integer vector of the number of results returned on a get_zen() query. Defaults to NULL.

Value

A list of class zengine.